+++ src/components/Identicon.jsx
@@ -17,7 +17,7 @@ export default class Identicon extends Component {
     address: PropTypes.string,
     anonymous: PropTypes.bool,
     classes: PropTypes.string,
-    size: PropTypes.oneOf(['tiny', 'small', 'medium', 'large'])
+    size: PropTypes.oneOf(['nano', 'tiny', 'small', 'medium', 'large'])
   }
 
   static defaultProps = {
@@ -66,6 +66,11 @@ export default class Identicon extends Component {
   }
 }
 const config = {
+  nano: {
+    size: '10px',
+    boxShadow:
+      'inset 0 1px 1px hsla(0,0%,100%,.1), inset 0 -1px 1px rgba(0,0,0,.1)'
+  },
   tiny: {
     size: '21px',
     boxShadow: