+++ src/components/Network/NodeInfo/NodeInfoBox.jsx
@@ -73,7 +73,7 @@ export default class NodeInfoBox extends Component {
     return (
       <div>
         <div className="looking-for-peers row-icon">
-          <FontAwesomeIcon icon="share-alt" />
+          <FontAwesomeIcon icon={faShareAlt} />
           {i18n.t('mist.nodeInfo.lookingForPeers')}
         </div>
       </div>
@@ -87,11 +87,11 @@ export default class NodeInfoBox extends Component {
     return (
       <div>
         <div className="peer-count row-icon">
-          <FontAwesomeIcon icon="users" />
+          <FontAwesomeIcon icon={faUsers} />
           {` ${connectedPeers} ${i18n.t('mist.nodeInfo.peers')}`}
         </div>
         <div className="sync-starting row-icon">
-          <FontAwesomeIcon icon="bolt" />
+          <FontAwesomeIcon icon={faBolt} />
           {i18n.t('mist.nodeInfo.syncStarting')}
         </div>
       </div>
@@ -111,15 +111,15 @@ export default class NodeInfoBox extends Component {
     return (
       <div>
         <div className="block-number row-icon">
-          <FontAwesomeIcon icon="layer-group" />
+          <FontAwesomeIcon icon={faLayerGroup} />
           {formattedCurrentBlock}
         </div>
         <div className="peer-count row-icon">
-          <FontAwesomeIcon icon="users" />
+          <FontAwesomeIcon icon={faUsers} />
           {` ${connectedPeers} ${i18n.t('mist.nodeInfo.peers')}`}
         </div>
         <div className="sync-progress row-icon">
-          <FontAwesomeIcon icon="cloud-download-alt" />
+          <FontAwesomeIcon icon={faCloudDownloadAlt} />
           <progress max="100" value={progress || 0} />
         </div>
       </div>
@@ -143,12 +143,12 @@ export default class NodeInfoBox extends Component {
           className="block-number row-icon"
           title={i18n.t('mist.nodeInfo.blockNumber')}
         >
-          <FontAwesomeIcon icon="layer-group" />
+          <FontAwesomeIcon icon={faLayerGroup} />
           {formattedBlockNumber}
         </div>
         {network !== 'private' && (
           <div className="peer-count row-icon">
-            <FontAwesomeIcon icon="users" />
+            <FontAwesomeIcon icon={faUsers} />
             {` ${connectedPeers} ${i18n.t('mist.nodeInfo.peers')}`}
           </div>
         )}
@@ -160,7 +160,7 @@ export default class NodeInfoBox extends Component {
           {
             // TODO: make this i8n compatible
           }
-          <FontAwesomeIcon icon="clock" />
+          <FontAwesomeIcon icon={faClock} />
           {diff < 120 ? `${diff} seconds` : `${Math.floor(diff / 60)} minutes`}
         </div>{' '}
       </div>
@@ -190,7 +190,7 @@ export default class NodeInfoBox extends Component {
           </div>
           <div>
             <div className="remote-loading row-icon">
-              <FontAwesomeIcon icon="bolt" />
+              <FontAwesomeIcon icon={faBolt} />
               {i18n.t('mist.nodeInfo.connecting')}
             </div>
           </div>
@@ -206,7 +206,7 @@ export default class NodeInfoBox extends Component {
           </span>
         </div>
         <div className="block-number row-icon">
-          <FontAwesomeIcon icon="layer-group" />
+          <FontAwesomeIcon icon={faLayerGroup} />
           {formattedBlockNumber}
         </div>
         <div
@@ -217,7 +217,7 @@ export default class NodeInfoBox extends Component {
           {
             // TODO: make this i8n compatible
           }
-          <FontAwesomeIcon icon="clock" />
+          <FontAwesomeIcon icon={faClock} />
           {diff < 120 ? `${diff} seconds` : `${Math.floor(diff / 60)} minutes`}
         </div>
       </div>