+++ src/components/Widgets/Form/RadioButton.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Radio = ({ children, onClick }) => (
+  <input type="text" className="input-field input-radio" />
+)
+
+Radio.displayName = 'RadioButton'
+
+export default Radio