+++ src/components/Widgets/Form/Select.jsx
@@ -1,0 +1,7 @@
+import Select from 'react-select'
+
+const CustomSelect = Select
+
+CustomSelect.displayName = 'Select'
+
+export default CustomSelect