+++ src/components/Widgets/Form/TextArea.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const TextArea = ({ children, onClick }) => (
+  <textarea className="input-field input-textarea" />
+)
+
+TextArea.displayName = 'TextArea'
+
+export default TextArea