+++ src/stories/3.Tx.stories.jsx
@@ -4,6 +4,10 @@ import FeeSelector from '../components/Tx/SendTx/FeeSelector'
 import TxHistory from '../components/Tx/History'
 import FormSubmitTx from '../components/Tx/SendTx/FormSubmitTx'
 import TxDescription from '../components/Tx/SendTx/TxDescription'
+import DeployContract from '../components/Tx/SendTx/TxDescription/DeployContract'
+import TokenTransfer from '../components/Tx/SendTx/TxDescription/TokenTransfer'
+import FunctionExecution from '../components/Tx/SendTx/TxDescription/FunctionExecution'
+import SendEther from '../components/Tx/SendTx/TxDescription/SendEther'
 import GasNotification from '../components/Tx/SendTx/GasNotification'
 import TxParties from '../components/Tx/SendTx/TxParties'
 import TxParty from '../components/Tx/SendTx/TxParty'
@@ -17,7 +21,11 @@ const dummyTx = {
   data: '',
   gasPrice: '0x9184e72a000', // 10000000000000
   value: '1000000000000000000',
-  params: [{ value: '0x4444444444444444444444444444444444444444' }]
+  params: [
+    { value: '0x4444444444444444444444444444444444444444' },
+    { value: '20000000000000000' }
+  ],
+  network: 'main'
 }
 
 storiesOf('Tx/Fee Selector', module).add('default ', () => <FeeSelector />)
@@ -27,9 +35,67 @@ storiesOf('Tx/Submit Form', module)
   .add('confirming', () => <FormSubmitTx unlocking />)
   .add('error', () => <FormSubmitTx error />)
 
-storiesOf('Tx/Description', module).add('default', () => {
-  return <TxDescription />
+storiesOf('Tx/Description', module)
+  .add('default', () => {
+    return <TxDescription {...dummyTx} />
+  })
+  .add('deploy contract', () => {
+    return <TxDescription {...dummyTx} isNewContract />
+  })
+  .add('transfer tokens', () => {
+    return (
+      <TxDescription
+        {...dummyTx}
+        executionFunction="transfer(address,uint256)"
+        token={{ symbol: 'LOL', decimals: 18 }}
+        toIsContract
+      />
+    )
+  })
+  .add('execute function', () => {
+    return (
+      <TxDescription
+        {...dummyTx}
+        executionFunction="approve(uint256)"
+        toIsContract
+      />
+    )
+  })
+
+storiesOf('Tx/Description/DeployContract', module).add('with data', () => {
+  return <DeployContract data={'a'.repeat(500)} />
 })
+
+storiesOf('Tx/Description/TokenTransfer', module)
+  .add('with token data', () => {
+    return (
+      <TokenTransfer
+        params={[{ value: '?' }, { value: '800000000000000' }]}
+        token={{ symbol: 'LOL', decimals: 18 }}
+      />
+    )
+  })
+  .add('without token data', () => {
+    return (
+      <TokenTransfer params={[{ value: '?' }, { value: '800000000000000' }]} />
+    )
+  })
+
+storiesOf('Tx/Description/SendEther', module)
+  .add('default', () => {
+    return <SendEther value="0.03" valueInUSD="3" network="main" />
+  })
+  .add('transfer', () => {
+    return <SendEther value="0.03" valueInUSD="3" network="rinkeby" />
+  })
+
+storiesOf('Tx/Description/FunctionExecution', module)
+  .add('default', () => {
+    return <FunctionExecution />
+  })
+  .add('transfer', () => {
+    return <FunctionExecution executionFunction="transfer(uint256,address)" />
+  })
 
 storiesOf('Tx/TxParty', module)
   .add('origin', () => {